pyproject.toml
[project]
name = "loopr"
version = "0.1.0"
description = "Classification of Carbon Fiber Defects"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "torchaudio>=2.5.0",
    "opencv-python>=4.10.0",
    "Pillow>=11.0.0",
    "numpy>=2.0.0",
    "matplotlib>=3.10.8",
    "scipy>=1.17.1",
    "scikit-learn>=1.8.0",
    "albumentationsx[headless]>=2.0.20",
    "hydra-core>=1.3.2",
    "tqdm>=4.67.3",
    "ray[serve]>=2.54.0",
    "fastapi[standard]>=0.135.1",
]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true


[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
pythonpath = [ "src"]

[tool.fastapi]
entrypoint = "loopr.serve_model:app"
